Refactor patrol setup
Introduced PatrolSetup MonoBehaviour to configure patrol positions per enemy instance, storing them in the Enemy's stateContext dictionary. Removed patrol position fields from IdlePatrol ScriptableObject and updated logic to read from stateContext, allowing unique patrol paths for each enemy.

Not sure if this is the right way to do it, will investigate. This is for variables that have to be different between different objects with the same scriptable object.
For instance, two enemies with IdlePatrol will have the same SO. If I want different positions to patrol, I can't set that up in the SO, because then all enemies with that SO will have the same positions, making me do a different SO for each different position, something completely stupid. For now this is the fix, will
think of cleaner ways to do it.