Change collider check in states transitions
Due to the addition of too many colliders to the player that do not have the tag Player, the check that the enemy does to know if it has line of sight of the player was incorrect. Now checks the root of the collision hit.
More bugs in the pathfinding, in the combat test scene sometimes the enemy in unable to create a path and gets stucked.